* Add a new component for file_progress
